fix media query and nav bar
Nav bar now changes to a dropdown-menu for smaller screens.

Would want to amend this to be an onClick function instead of a hover pseudo-class to make it easier for mobile users, but since the FCC project is for HTML/CSS only I'll leave this functionality for a later project.